refactor: Remove dataset-specific code and legacy APIs from langres.data
This is an aggressive cleanup following the POC principle: "make it work
first before worrying about backwards compatibility."

Changes:
- Remove load_iteration_05_data() (dataset-specific function)
- Remove load_labeled_dedup_data_legacy() (convenience wrapper)
- Remove backward compatibility for variant_ids/variant_names fields
- Simplify stratified_dedup_split() to accept only LabeledDeduplicationDataset
- Consolidate test files (remove test_loaders_new.py, test_splitting_new.py)
- Rename deduplication_iteration_05.py → deduplication_example.py
- Update example to use generic loader with custom file names
- Remove all TODO comments related to cleanup
- Update public API exports in __init__.py

Result: Clean, minimal data loading API with no dataset coupling.
The library provides generic infrastructure; examples show how users
should adapt their data to the langres schema.

Net: -679 lines (removed legacy code and tests)
